requirements.txt
flask==2.0.1
gunicorn==20.1.0
llama-cpp-python>=0.2.38
huggingface_hub==0.19.4